+++ src/main/java/org/example/first_project/service/imp/EmployeeServiceImp.java
@@ -5,10 +5,12 @@ import org.example.first_project.dao.EmployeeDao;
 import org.example.first_project.model.dto.DepartmentDtoRes;
 import org.example.first_project.model.dto.EmployeeDtoReq;
 import org.example.first_project.model.dto.EmployeeDtoRes;
+import org.example.first_project.model.entity.Address;
 import org.example.first_project.model.entity.Employee;
 import org.example.first_project.model.mapper.AddressMapper;
 import org.example.first_project.model.mapper.DepartmentMapper;
 import org.example.first_project.model.mapper.EmployeeMapper;
+import org.example.first_project.service.AddressService;
 import org.example.first_project.service.DepartmentService;
 import org.example.first_project.service.EmployeeService;
 import org.springframework.stereotype.Service;
@@ -26,6 +28,8 @@ public class EmployeeServiceImp implements EmployeeService {
 
     private final DepartmentMapper departmentMapper;
 
+    private final AddressService addressService;
+
     @Override
     public List<Employee> getEmployees() {
         return employeeDao.getEmployees();
@@ -33,9 +37,10 @@ public class EmployeeServiceImp implements EmployeeService {
 
     @Override
     public EmployeeDtoRes getEmployeeById(Long id) {
-        Employee employee=employeeDao.getEmployeeById(id).orElse(new Employee());
+        Employee employee = employeeDao.getEmployeeById(id).orElse(new Employee());
         EmployeeDtoRes res = employeeMapper.toDtoRes(employee);
         res.setDepartment(departmentMapper.toRes(employee.getDepartment()));
+        res.setAddress(addressMapper.toDto(employee.getAddress()));
         return res;
     }
 
@@ -43,19 +48,19 @@ public class EmployeeServiceImp implements EmployeeService {
     public EmployeeDtoRes createEmployee(EmployeeDtoReq req) {
         Employee employee = employeeMapper.toEntity(req);
         if (req.getDepartment().getId() != 0) {
-            DepartmentDtoRes departmentseved = departmentService.getDepartmentById(req.getDepartment().getId());
+            DepartmentDtoRes departmentseved = departmentService.getDepartmentById(req.getDepartment().getId()).get();
             employee.setDepartment(departmentMapper.toEntity(departmentMapper.toReq(departmentseved)));
             System.out.println(departmentseved);
         } else {
             employee.setDepartment((departmentMapper.toEntity(req.getDepartment())));
         }
-        employee.setAddress(addressMapper.toEntity(req.getAddress()));
+        Address address = addressMapper.toEntity(req.getAddress());
         employee = employeeDao.createEmployee(employee);
-//        System.out.println(employee.getDepartment().getName());
-      //  System.out.println(employee.getAddress().getId());
+        address.setEmployee(employee);
+        addressService.createAddress(address);
         EmployeeDtoRes res = employeeMapper.toDtoRes(employee);
         res.setDepartment(departmentMapper.toRes(employee.getDepartment()));
-        res.setAddress(addressMapper.toDto(employee.getAddress()));
+        res.setAddress(addressMapper.toDto(address));
         return res;
 
     }
@@ -64,16 +69,32 @@ public class EmployeeServiceImp implements EmployeeService {
     public EmployeeDtoRes updateEmployee(EmployeeDtoReq req) {
         Employee employee = employeeMapper.toEntity(req);
         employee.setDepartment(departmentMapper.toEntity(req.getDepartment()));
-        System.out.println(employee.getName());
+        employee.setAddress(addressMapper.toEntity(req.getAddress()));
         employee = employeeDao.updateEmployee(employee);
         EmployeeDtoRes res = employeeMapper.toDtoRes(employee);
+        System.out.println(employee.getAddress().getId());
         res.setDepartment(departmentMapper.toRes(employee.getDepartment()));
+        res.setAddress(addressMapper.toDto(employee.getAddress()));
         return res;
 
     }
 
     @Override
+    public boolean isExist(Long id) {
+        return employeeDao.isExistById(id);
+    }
+
+    @Override
     public boolean deleteEmployeeById(long id) {
-        return employeeDao.deleteDepartmentById(id);
+        if (this.isExist(id)) {
+            try {
+                employeeDao.deleteDepartmentById(id);
+                return true;
+            } catch (Exception e) {
+                return false;
+            }
+        }
+        return false;
+
     }
 }
